refactor: Main -> Home 페이지 이름 변경
