+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import Main from './pages/Main';
+import Home from './pages/Home';
 import SignUp from './pages/SignUp';
 import GlobalStyle from './styles/GlobalStyle';
 import GlobalFonts from './styles/GlobalFont';
@@ -9,7 +9,7 @@ function App() {
   const router = createBrowserRouter([
     {
       path: '/',
-      element: <Main />,
+      element: <Home />,
     },
     {
       path: '/signUp',
